[web] move some vector functions into vector_math.dart
