+++ lib/web_ui/lib/src/engine/canvaskit/canvaskit_canvas.dart
@@ -6,8 +6,8 @@ import 'dart:typed_data';
 
 import 'package:ui/ui.dart' as ui;
 
-import '../../engine.dart' show toMatrix32;
 import '../validators.dart';
+import '../vector_math.dart';
 import 'canvas.dart';
 import 'canvaskit_api.dart';
 import 'image.dart';
