+++ lib/web_ui/lib/src/engine/html/canvas.dart
@@ -7,10 +7,10 @@ import 'dart:typed_data';
 
 import 'package:ui/ui.dart' as ui;
 
-import '../../engine.dart' show toMatrix32;
 import '../picture.dart';
 import '../util.dart';
 import '../validators.dart';
+import '../vector_math.dart';
 import 'painting.dart';
 import 'recording_canvas.dart';
 import 'render_vertices.dart';
