+++ lib/web_ui/lib/src/engine/html/scene_builder.dart
@@ -6,7 +6,7 @@ import 'dart:typed_data';
 
 import 'package:ui/ui.dart' as ui;
 
-import '../../engine.dart' show kProfileApplyFrame, kProfilePrerollFrame, toMatrix32;
+import '../../engine.dart' show kProfileApplyFrame, kProfilePrerollFrame;
 import '../picture.dart';
 import '../profiler.dart';
 import '../util.dart';
