+++ lib/web_ui/lib/src/engine/initialization.dart
@@ -318,39 +318,3 @@ void _addUrlStrategyListener() {
 /// rendering of PlatformViews into the web app.
 // TODO(dit): How to make this overridable from tests?
 final PlatformViewManager platformViewManager = PlatformViewManager();
-
-// TODO(yjbanov): this does not belong here.
-//                https://github.com/flutter/flutter/issues/100394
-
-/// Converts a matrix represented using [Float64List] to one represented using
-/// [Float32List].
-///
-/// 32-bit precision is sufficient because Flutter Engine itself (as well as
-/// Skia) use 32-bit precision under the hood anyway.
-///
-/// 32-bit matrices require 2x less memory and in V8 they are allocated on the
-/// JavaScript heap, thus avoiding a malloc.
-///
-/// See also:
-/// * https://bugs.chromium.org/p/v8/issues/detail?id=9199
-/// * https://bugs.chromium.org/p/v8/issues/detail?id=2022
-Float32List toMatrix32(Float64List matrix64) {
-  final Float32List matrix32 = Float32List(16);
-  matrix32[15] = matrix64[15];
-  matrix32[14] = matrix64[14];
-  matrix32[13] = matrix64[13];
-  matrix32[12] = matrix64[12];
-  matrix32[11] = matrix64[11];
-  matrix32[10] = matrix64[10];
-  matrix32[9] = matrix64[9];
-  matrix32[8] = matrix64[8];
-  matrix32[7] = matrix64[7];
-  matrix32[6] = matrix64[6];
-  matrix32[5] = matrix64[5];
-  matrix32[4] = matrix64[4];
-  matrix32[3] = matrix64[3];
-  matrix32[2] = matrix64[2];
-  matrix32[1] = matrix64[1];
-  matrix32[0] = matrix64[0];
-  return matrix32;
-}
